NOTIF-138 Move static inner classes out of future JPA entities
