+++ src/main/java/com/redhat/cloud/notifications/models/Endpoint.java
@@ -15,19 +15,6 @@ import java.util.UUID;
 
 public class Endpoint {
 
-    // Add new values to the bottom of the enum
-    // The ordinal order must remain the same as is used internally
-    // Update test com.redhat.cloud.notifications.models.TestEndpointType to reflect any new enum value
-    @Schema(enumeration = { "webhook", "email_subscription", "default" })
-    public enum EndpointType {
-        @JsonProperty("webhook")
-        WEBHOOK, // 0
-        @JsonProperty("email_subscription")
-        EMAIL_SUBSCRIPTION, // 1
-        @JsonProperty("default")
-        DEFAULT // 2
-    }
-
     private UUID id; // Should be UUID
 
     @JsonIgnore
