+++ src/main/java/com/redhat/cloud/notifications/models/HttpType.java
@@ -1,0 +1,5 @@
+package com.redhat.cloud.notifications.models;
+
+public enum HttpType {
+    GET, POST, PUT
+}
